Multiple changes to support more complex structs
The biggest issues being utilizing provided type definitions as type
parameters within the context of other provided types. Initial run is
swapping out Option for a new type IsSet<'a> that has an IIsSet
interface that will allow access to HasValue/Value in a non-generic
context. Also switched to public fields since setters on properties with
a generated type as a type definition would also fail.